back-end/requirements.txt
flask
Flask-SQLAlchemy
numpy
python-dotenv